update all the things
