+++ content/index.mdx
@@ -1,10 +1,6 @@
----
-layout: mdx
----
+# HashiCorp Web Engineering Challenge
 
-import Link from 'next/link'
-
-## HashiCorp Marketing Webdev Challenge
+Hi! Thanks so much for interviewing with us 💖. If you are reading this, it means that we like you, and I hope you like us too! In this next step, you get an opportunity to see how we work, and we get the same opportunity for you. Let's get right to it.
 
 The task is to build a "people directory" page. This page should allow visitor to search by name for a person. If this were a real page, we'd want to search all 317 person records in our CMS, but for the purposes of this excercise, limit your results to 100. This is the maximum number of results that our CMS API will return in a single page.
 
@@ -17,24 +13,25 @@ Use [this Figma file](https://www.figma.com/file/xGicP4qkXbMhte4LAYxC4X/Untitled
   allowFullScreen
 ></iframe>
 
-Treat this the same way you would if you started working with us and this was your first day -- that is, feel free to ask as many questions as you want. We will set up a slack channel for you to ask questions and interact with us. This is quite similar to the type of task we would assign a new hire on their first day, in fact.
+There are definitely some parts of this that are a little ambiguous, which is how real world projects always work. Treat this the same way you would if you started working with us and this was your first day — this is quite similar to the type of task we would assign a new hire on their first day, in fact. We will set up a slack channel for you to ask questions, get clarification, and collaborate with us in general, and we strongly encourage you to take advantage of this.
 
-We expect this to be done within a day or two's worth of full-time work - about 8 - 14 hours. But obviously you have other committments and so we generally look at about week for candidates to complete the challenge. If you're approaching 14 hours (of total time spent), just stop where you are and we'll complete the evaluation.
+We expect this to be done within a day or two's worth of full-time work - about 8 - 14 hours. We know that you likely have other commitments, and so we generally look at about week for candidates to complete the challenge. If you're approaching 14 hours of total time spent, just stop where you are and we'll complete the evaluation based on what you have done.
 
 ### Fetching Data
 
-This project uses [Next.js](https://nextjs.org), a react-based framework that can generate static or dynamic websites. We have a tool called `nextjs-scripts` layered into the config that provides several utilities, especially helpful for loading remote data.
+This project uses [Next.js](https://nextjs.org), a react-based framework that generates performant websites that blur the lines between static and dynamically rendered pages. We have an internal tool called `nextjs-scripts` layered into the config that provides several utilities, and is especially helpful for loading remote data.
 
 The data for this project can be fetched from our CMS, [DatoCMS](https://www.datocms.com/), `nextjs-scripts` exports a function that allows you to easily fetch the data. An example might look something like this:
 
 ```js
 import rivetQuery from '@hashicorp/nextjs-scripts/dato/client'
+// nextjs-scripts configures ".graphql" files to be imported as plain text
 import someQuery from './query.graphql'
 
 rivetQuery({ query: someQuery }).then(console.log)
 ```
 
-Next.js [conventions](https://nextjs.org/learn/basics/fetching-data-for-pages), however, dictate that an async function called `getStaticProps` can be exported along with any component within the `pages` directory. This function will be executed before the component mounts, passing the returned object into the component as props. So, within this project, your initial data fetch might look something like this:
+Next.js [conventions](https://nextjs.org/learn/basics/fetching-data-for-pages) dictate that an async function called `getStaticProps` can be exported along with any component within the `pages` directory. This function will be executed before the component mounts, passing the returned object into the component as props. So, within this project, your initial data fetch might look something like this:
 
 ```js
 import rivetQuery from '@hashicorp/nextjs-scripts/dato/client'
@@ -70,8 +67,6 @@ You may have also discovered that we have a library of components under the `@ha
 
 All images on the page can be retrieved from the CMS via URLs from the graphql query. And we expect the page to be styled responsively so it looks great on both mobile and desktop.
 
-<Link href="/people">
-  <a>&raquo; Here's the page!</a>
-</Link>
+<Button href="/people">Here's the page! 👉</Button>
 
 ![excited bun](https://media.giphy.com/media/cMnt7i2RykmpW/giphy.gif)
