--- layouts/style.module.css
@@ -1,10 +1,0 @@
-.intro {
-  max-width: 1000px;
-  margin: 0 auto;
-
-  & img {
-    width: 50%;
-    display: block;
-    margin: 50px auto;
-  }
-}
