+++ next-env.d.ts
@@ -1,0 +1,4 @@
+/// <reference types="next" />
+/// <reference types="next/types/global" />
+
+declare module '*.graphql'
