+++ package.json
@@ -4,29 +4,25 @@
   "version": "0.1.0",
   "author": "HashiCorp",
   "dependencies": {
-    "@hashicorp/nextjs-scripts": "13.2.0",
-    "@hashicorp/react-content": "5.2.1",
+    "@hashicorp/mktg-global-styles": "^3.1.0",
+    "@hashicorp/nextjs-scripts": "19.0.3",
+    "@hashicorp/react-content": "7.0.1",
     "@hashicorp/react-email-subscribe": "2.1.8",
     "@hashicorp/react-footer": "3.3.4",
-    "@hashicorp/react-global-styles": "4.6.1",
     "@hashicorp/react-nav": "3.0.4",
     "@hashicorp/react-secondary-nav": "2.1.1",
-    "@hashicorp/react-section-header": "3.0.0",
-    "faker": "5.1.0",
-    "imagemin-mozjpeg": "8.0.0",
-    "imagemin-optipng": "7.1.0",
-    "imagemin-svgo": "7.1.0",
-    "marked": "0.7.0",
-    "next": "9.5.2",
-    "nprogress": "^0.2.0",
+    "@hashicorp/react-section-header": "5.0.2",
+    "faker": "5.5.3",
+    "next": "10.2.3",
+    "next-mdx-remote": "^3.0.2",
     "react": "16.14.0",
-    "react-dom": "16.14.0",
-    "slugify": "1.4.5"
+    "react-dom": "16.14.0"
   },
   "devDependencies": {
+    "@types/react": "^17.0.14",
     "husky": "4.2.5",
     "inquirer": "^7.0.0",
-    "prettier": "2.1.0"
+    "prettier": "2.3.2"
   },
   "husky": {
     "hooks": {
@@ -36,7 +32,6 @@
   "main": "index.js",
   "scripts": {
     "build": "next build",
-    "prod": "NODE_ENV=production next build && next start",
     "format": "next-hashicorp format",
     "lint": "next-hashicorp lint",
     "start": "next dev",
