--- pages/_app.js
@@ -1,35 +1,0 @@
-import React from 'react'
-import './style.css'
-import '@hashicorp/nextjs-scripts/lib/nprogress/style.css'
-
-import NProgress from '@hashicorp/nextjs-scripts/lib/nprogress'
-import Router from 'next/router'
-import Layout from '../layouts/hashicorp'
-
-NProgress({ Router })
-
-function App({ Component, pageProps }) {
-  const AppLayout = Component.layout ? Layout : React.Fragment
-  return (
-    <AppLayout>
-      <Component {...pageProps} />
-    </AppLayout>
-  )
-}
-
-App.getInitialProps = async ({ Component, ctx }) => {
-  let pageProps = {}
-  if (Component.getInitialProps) {
-    pageProps = await Component.getInitialProps(ctx)
-  } else if (Component.isMDXComponent) {
-    // fix for https://github.com/mdx-js/mdx/issues/382
-    const mdxLayoutComponent = Component({}).props.originalType
-    if (mdxLayoutComponent.getInitialProps) {
-      pageProps = await mdxLayoutComponent.getInitialProps(ctx)
-    }
-  }
-
-  return { pageProps }
-}
-
-export default App
