--- pages/_document.js
@@ -1,26 +1,0 @@
-import Document, { Head, Main, NextScript } from 'next/document'
-
-export default class MyDocument extends Document {
-  static async getInitialProps(ctx) {
-    const initialProps = await Document.getInitialProps(ctx)
-    return { ...initialProps }
-  }
-
-  render() {
-    return (
-      <html lang="en">
-        <Head>
-          <link rel="stylesheet" type="text/css" href="/css/nprogress.css" />
-          <link
-            href="https://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700&display=swap"
-            rel="stylesheet"
-          />
-        </Head>
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </html>
-    )
-  }
-}
