+++ pages/index.tsx
@@ -1,0 +1,4 @@
+import IndexPage, { getStaticProps as _getStaticProps } from './home'
+
+export default IndexPage
+export const getStaticProps = _getStaticProps
