+++ pages/people/index.tsx
