+++ pages/style.css
@@ -1,4 +1,6 @@
-@import '~@hashicorp/react-global-styles/style.css';
+@import '~@hashicorp/mktg-global-styles/style.css';
+@import '~@hashicorp/nextjs-scripts/prism/style.css';
+
 @import '~@hashicorp/react-nav/dist/style.css';
 @import '~@hashicorp/react-footer/dist/style.css';
 @import '~@hashicorp/react-newsletter-signup-form/dist/style.css';
