--- prettier.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  ...require('@hashicorp/nextjs-scripts/prettier.config.js'),
-  /* Specify overrides here */
-}
