+++ pull_request_template.md
@@ -1,5 +1,5 @@
 Please **do not submit a pull request to this repository!**
 
-Instead, use the "use this template" button via github to create your own copy of the repo, grant access to `kylemac`, `jescalan`, and `mwickett`, and open a pull request against your own copy for review.
+Instead, use the "use this template" button via github to create your own copy of the repo, grant admin access to `kylemac`, `jescalan`, and `mwickett`, and open a pull request against your own copy for review.
 
 We look forward to checking out your code!
